[Style] Extensions 구조 변경
