+++ Extensions/Color+Extensions.swift
@@ -1,0 +1,29 @@
+//
+//  Color+Extensions.swift
+//  Record
+//
+//  Created by 이지원 on 2022/10/05.
+//
+
+import SwiftUI
+
+extension Color {
+    // 피그마에 등록된 색상 보시고, 사용하시면 됩니다!
+    // foreground(.titleBlack) 이런식으로 사용하시면 돼요!
+    
+    //title 색상
+    static let titleBlack = Color("black")
+    static let titleDarkgray = Color("darkgray")
+    static let titleGray = Color("gray")
+    static let titleLightgray = Color("lightgray")
+    
+    //point color
+    static let pointBlue = Color("blue")
+    static let pointOrange = Color("orange")
+    static let pointYellow = Color("yellow")
+    
+    //etc
+    static let background = Color("background")
+    static let cdplayer = Color("cdplayer")
+
+}
