+++ Extensions/View+Extensions.swift
@@ -1,0 +1,21 @@
+//
+//  View+Extensions.swift
+//  Record
+//
+//  Created by 이지원 on 2022/10/05.
+//
+
+import SwiftUI
+
+extension View {
+    func placeholder<Content: View>(
+        when shouldShow: Bool,
+        alignment: Alignment = .leading,
+        @ViewBuilder placeholder: () -> Content) -> some View {
+
+        ZStack(alignment: alignment) {
+            placeholder().opacity(shouldShow ? 1 : 0)
+            self
+        }
+    }
+}
